embedding_compare/requirements.txt
google-generativeai
numpy
dotenv
scipy